Implement std.functional.stringLambda.
stringLambda translates a string lambda into a function. The function can
have an arbitrary number of args. These are assumed to use the letters
a-z, but the caller can provide custom arg names. If the string lambda
uses > 26 args, the caller can use args[n] to access the nth arg.

This is being implemented to support multi-arg std.algorithm.each.

Also replaces $(D ...) with `...` in stringLambda docs.

Includes a number of ideas from @tsbockman to:
    - reduce CTFE load
    - enforce matching argument count
    - keep template internals private

naryFunAliases is implemented as a private template _outside_ of to
reduce CTFE work by caching the result when fun differs but paramNames
are the same. This ensures the aliases are computed only once for each
unique set of paramNames or each value of argc.